refactor user related helper functions to its own util file
